[Refactor] Meatballs Menu 코드 분리
