+++ client/src/components/MeatballsMenu.jsx
@@ -1,11 +1,16 @@
 import { useState } from 'react';
 import styled from 'styled-components';
-import { BsTrashFill } from 'react-icons/bs';
+import { BsThreeDots, BsTrashFill } from 'react-icons/bs';
 import { HiOutlinePencil } from 'react-icons/hi';
 import Modal from './Modal/Modal';
 
 function MeatballsMenu() {
+  const [isEdit, setIsEdit] = useState(false);
   const [showModal, setShowModal] = useState(false);
+
+  const onEditHandler = () => {
+    setIsEdit(!isEdit);
+  };
 
   const showModalHandler = () => {
     setShowModal(!showModal);
@@ -13,25 +18,50 @@ function MeatballsMenu() {
 
   return (
     <MenuContainer>
-      <MenuButton>
-        <EditIcon />
-        수정하기
-      </MenuButton>
-      <MenuButton onClick={showModalHandler}>
-        <DeleteIcon />
-        삭제하기
-      </MenuButton>
-      <Modal
-        showModal={showModal}
-        handleCloseModal={showModalHandler}
-        title="게시글 삭제"
-        content="정말 게시글을 삭제하시겠습니까?"
-      />
+      <MeatballsButton onClick={onEditHandler}>
+        <MenuIcon />
+      </MeatballsButton>
+      {isEdit && (
+        <MenuWrapper>
+          <MenuButton>
+            <EditIcon />
+            수정하기
+          </MenuButton>
+          <MenuButton onClick={showModalHandler}>
+            <DeleteIcon />
+            삭제하기
+          </MenuButton>
+          <Modal
+            showModal={showModal}
+            handleCloseModal={showModalHandler}
+            title="게시글 삭제"
+            content="정말 게시글을 삭제하시겠습니까?"
+          />
+        </MenuWrapper>
+      )}
     </MenuContainer>
   );
 }
 
 const MenuContainer = styled.div`
+  display: flex;
+  position: relative;
+`;
+
+const MeatballsButton = styled.button`
+  padding: 0;
+  border: none;
+  background: none;
+`;
+
+const MenuIcon = styled(BsThreeDots)`
+  width: 20px;
+  height: 20px;
+  color: var(--font-color);
+  cursor: pointer;
+`;
+
+const MenuWrapper = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: space-evenly;
